// src/app/models/Task.ts
export class Task {
    id: number;
    read: boolean;
    assign_to: string;
    task_date: string;
    task_description: string;
    application_id: string;
    created_by: string;
    modified_by: string;
    delete: boolean;
    assign_fname : string;
    assign_lname : string;
    assign_by_fname : string;
    assign_by_lname : string;
    status : string; 
}